fixup! [PAL/Linux-SGX] Add EDMM support

+++ pal/src/host/linux-sgx/pal_files.c
@@ -274,7 +274,7 @@ static int file_map(PAL_HANDLE handle, void* addr, pal_prot_flags_t prot, uint64
 
     if (g_pal_linuxsgx_state.edmm_enabled) {
         /* Enclave pages will be written to below, so we must add W permission. */
-        ret = sgx_edmm_add_pages((uintptr_t)addr, size / PAGE_SIZE,
+        ret = sgx_edmm_add_pages((uint64_t)addr, size / PAGE_SIZE,
                                  PAL_TO_SGX_PROT(prot | PAL_PROT_WRITE));
         if (ret < 0) {
             return ret;
